example_project/requirements.txt
aiogram-deta @ git+https://github.com/aiogram/deta@main
